+++ lottery-data/src/main/java/de/alpharogroup/lottery/enums/LotteryAlgorithm.java
@@ -1,0 +1,38 @@
+/**
+ * Commercial License
+ *
+ * Copyright (C) 2015 Asterios Raptis - All Rights Reserved
+ *
+ * Proprietary and confidential
+ *
+ * Unauthorized copying of this software and its files,
+ * via any medium is strictly prohibited
+ *
+ * Written by Asterios Raptis
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package de.alpharogroup.lottery.enums;
+
+/**
+ * The enum {@link LotteryAlgorithm} holds algorithm types for drawing lottery numbers
+ */
+public enum LotteryAlgorithm
+{
+
+	/** The default algorithm */
+	DEFAULT,
+
+	/** The map algorithm */
+	MAP,
+
+	/** The set algorithm */
+	SET;
+
+}